bug 1564452: move storage code around
This starts a tecken/libstorage.py module with the storage base classes
and errors. This moves the storage code (which is aws s3-centric) to
tecken/ext/s3/storage.py.

This updates all the tests and code to match the new locations.
