+++ tecken/libdockerflow.py
@@ -7,8 +7,8 @@ from pathlib import Path
 
 from django.core import checks
 
-from tecken.storage import StorageError
 from tecken.base.symbolstorage import symbol_storage
+from tecken.libstorage import StorageError
 
 
 def check_storage_urls(app_configs, **kwargs):
